move test files to relevant subfolders
