+++ tests/class_info/test_class_matches.py
