+++ tests/pairedomics/test_downloader.py
@@ -10,7 +10,7 @@ from nplinker import utils
 from nplinker.pairedomics.downloader import Downloader
 from nplinker.pairedomics.downloader import _generate_gnps_download_url
 from nplinker.pairedomics.downloader import _execute_download
-from . import DATA_DIR
+from .. import DATA_DIR
 
 
 @pytest.fixture
@@ -80,7 +80,7 @@ def test_download_gnps_data(tmp_path):
     gnps_task_id = "c22f44b14a3d450eb836d607cb9521bb"
     sut = Downloader("MSV000079284", local_cache=tmp_path / 'actual')
     actual = sut._load_gnps_data(gnps_task_id)
-    
+
     expected = zipfile.ZipFile(DATA_DIR / "ProteoSAFe-METABOLOMICS-SNETS-c22f44b1-download_clustered_spectra.zip")
 
     actual.extract("networkedges_selfloop/6da5be36f5b14e878860167fa07004d6.pairsinfo", tmp_path / "actual")
@@ -101,4 +101,3 @@ def test_extract_metabolomics_data(tmp_path):
     assert (Path(sut.project_file_cache) / "networkedges_selfloop/6da5be36f5b14e878860167fa07004d6.pairsinfo").is_file()
     assert (Path(sut.project_file_cache) / "clusterinfosummarygroup_attributes_withIDs_withcomponentID/d69356c8e5044c2a9fef3dd2a2f991e1.tsv").is_file()
     assert (Path(sut.project_file_cache) / "spectra/METABOLOMICS-SNETS-c22f44b1-download_clustered_spectra-main.mgf").is_file()
-
