+++ tests/test_annotations.py
@@ -1,6 +1,6 @@
 from nplinker.annotations import load_annotations
 
-from .test_metabolomics import spec_dict
+from .metabolomics.test_metabolomics import spec_dict
 from . import DATA_DIR
